refactor: update utils
Reference: #10, #14
