+++ src/lib/utils/index.ts
@@ -4,4 +4,3 @@ export * from './error-handler';
 export * from './files';
 export * from './jwt';
 export * from './middleware';
-export * from './types';
