+++ src/lib/utils/middleware.ts
@@ -1,52 +1,45 @@
 import { GraphQLError } from 'graphql';
 import { prisma } from '../../client';
+import { MESSAGES } from '../../config';
+import { ExpressRequest } from '../../types';
 import { verifyJwt } from './jwt';
-import { ExpressRequest } from './types';
 
 export const getAuthUser = async (req: ExpressRequest) => {
-  try {
-    let message = 'Invalid User: This user is not logged in';
-    const token = req.get('Authorization')?.split(' ')[1] || req.cookies?.token;
+  const token = req.get('Authorization')?.split(' ')[1] || req.cookies?.jwt;
 
-    if (!token) {
-      throw new GraphQLError(message, {
-        extensions: {
-          code: 'UNAUTHENTICATED',
-          http: { status: 401 },
-        },
-      });
-    }
-
-    message = 'Invalid Token: No valid keys or signatures';
-    const decoded = verifyJwt(token);
-    if (!decoded) {
-      throw new GraphQLError(message, {
-        extensions: {
-          code: 'UNAUTHENTICATED',
-          http: { status: 401 },
-        },
-      });
-    }
-
-    message = 'Invalid Token: This token or session has expired';
-    const user = await prisma.user.findUnique({
-      where: {
-        id: decoded?.sub,
+  if (!token) {
+    throw new GraphQLError(MESSAGES.SESSION_UNAUTHENTICATED, {
+      extensions: {
+        code: 'UNAUTHENTICATED',
+        http: { status: 401 },
       },
     });
-
-    if (!user) {
-      throw new GraphQLError(message, {
-        extensions: {
-          code: 'FORBIDDEN',
-          http: { status: 403 },
-        },
-      });
-    }
-
-    return user;
-  } catch (error) {
-    console.log(error);
-    return null;
   }
+
+  const decoded = verifyJwt(token);
+  if (!decoded) {
+    throw new GraphQLError(MESSAGES.SESSION_INVALID_TOKEN, {
+      extensions: {
+        code: 'UNAUTHENTICATED',
+        http: { status: 401 },
+      },
+    });
+  }
+
+  const user = await prisma.user.findUnique({
+    where: {
+      id: decoded?.sub,
+    },
+  });
+
+  if (!user) {
+    throw new GraphQLError(MESSAGES.SESSION_EXPIRED, {
+      extensions: {
+        code: 'FORBIDDEN',
+        http: { status: 403 },
+      },
+    });
+  }
+
+  return user;
 };
